Modularize L10N JSON File - UDX & Widget
