+++ gui/src/components/udx/udx.json
@@ -1,0 +1,120 @@
+{
+    "en": {
+        "settings": {
+            "dataNetworkIPv4": "Data network settings: IPv4",
+            "forManualOption": "For Manual option, you must configure Virtual IP address and values for Node 0 and Node 1.",
+            "mandatoryField": "Fields marked with * are mandatory.",
+            "DNSResolverSettings": "DNS resolver settings",
+            "managementSettingIPv4": "Management network settings: IPv4",
+            "managementSettingIPv4Msg": "Configure the Virtual IP address for managing this system.",
+            "notificationSettingsEmail": "Notification settings: Email",
+            "notificationSettingsEmailMsg": "Allows you to configure email servers for notifications.",
+            "NTPSettings": "Network time protocol (NTP) settings",
+            "SSLCertificateUpload": "SSL certificate upload",
+            "SSLCertificateUploadMsg": "By default, the system uses the SSL certificate provided by Seagate. To use a different SSL certificate, click Choose File to browse and select the appropriate SSL Certificate file. Click Upload certificate to upload the selected SSL certificate.",
+            "UDX": "UDX",
+            "name": "Name",
+            "product ID": "Product ID",
+            "serialNumber": "Serial number",
+            "type": "Type",
+            "UUID": "UUID",
+            "vendorID": "Vendor ID"
+        },
+        "udx-registration": {
+            "udx-registration": "Lyve Pilot Registration",
+            "registration-token": "Your registration token",
+            "udx-url-required": "URL is required",
+            "invalid-url": "Invalid URL",
+            "s3-details": "Account details",
+            "S3BucketDetails": "Bucket details",
+            "accountname-required": "Account name is required",
+            "invalid-name": "Invalid account name",
+            "email": "Email id*",
+            "email-reqiured": "Email id is required",
+            "invalid-email": "Invalid email id",
+            "password-required": "Password is required",
+            "invalid-password": "Invalid password",
+            "confirm-pass": "Confirm password*",
+            "password-match": "Passwords do not match",
+            "iam-details": "IAM User details",
+            "username-required": "Username is required",
+            "invalid-user": "Invalid username",
+            "bucket-required": "Bucket name is required",
+            "invalid-bucketname": "Invalid bucket name",
+            "iagreetext": "I agree to the following",
+            "firstcheck-text": "A dedicated Lyve Pilot Bucket will be created in my account.",
+            "secondcheck-text": "A Lyve Pilot IAM User will be created having complete access only to the Lyve Pilot Bucket, and not to other data.",
+            "register-btn": "Register",
+            "clear": "Clear",
+            "s3AccountKey": "S3 Account Access Key",
+            "s3AccountSecretKey": "S3 Account Secret Key",
+            "userAccessKey": "IAM User Access Key",
+            "userAccessSecretKey": "IAM User Secret Key",
+            "created": "Access key created",
+            "access_key": "Access key",
+            "secret_key": "Secret key",
+            "iam_access_key": "IAM Access key",
+            "iam_secret_key": "IAM Secret key",
+            "bucket_name":"Bucket Name"
+        }
+    },
+    "ru": {
+        "settings": {
+            "dataNetworkIPv4": "Data network settings: IPv4",
+            "forManualOption": "For Manual option, you must configure Virtual IP address and values for Node 0 and Node 1.",
+            "mandatoryField": "Fields marked with * are mandatory.",
+            "DNSResolverSettings": "DNS resolver settings",
+            "managementSettingIPv4": "Management network settings: IPv4",
+            "managementSettingIPv4Msg": "Configure the Virtual IP address for managing this system.",
+            "notificationSettingsEmail": "Notification settings: Email",
+            "notificationSettingsEmailMsg": "Allows you to configure email servers for notifications.",
+            "NTPSettings": "Network time protocol (NTP) settings",
+            "SSLCertificateUpload": "SSL certificate upload",
+            "SSLCertificateUploadMsg": "By default, the system uses the SSL certificate provided by Seagate. To use a different SSL certificate, click Choose File to browse and select the appropriate SSL Certificate file. Click Upload certificate to upload the selected SSL certificate.",
+            "UDX": "UDX",
+            "name": "Name",
+            "product ID": "Product ID",
+            "serialNumber": "Serial number",
+            "type": "Type",
+            "UUID": "UUID",
+            "vendorID": "Vendor ID"
+        },
+        "udx-registration": {
+            "udx-registration": "Lyve Pilot Registration",
+            "registration-token": "Your registration token",
+            "udx-url-required": "URL is required",
+            "invalid-url": "Invalid URL",
+            "s3-details": "Account details",
+            "S3BucketDetails": "Bucket details",
+            "accountname-required": "Account name is required",
+            "invalid-name": "Invalid account name",
+            "email": "Email id*",
+            "email-reqiured": "Email id is required",
+            "invalid-email": "Invalid email id",
+            "password-required": "Password is required",
+            "invalid-password": "Invalid password",
+            "confirm-pass": "Confirm password*",
+            "password-match": "Passwords do not match",
+            "iam-details": "IAM User details",
+            "username-required": "Username is required",
+            "invalid-user": "Invalid username",
+            "bucket-required": "Bucket name is required",
+            "invalid-bucketname": "Invalid bucket name",
+            "iagreetext": "I agree to the following",
+            "firstcheck-text": "A dedicated Lyve Pilot Bucket will be created in my account.",
+            "secondcheck-text": "A Lyve Pilot IAM User will be created having complete access only to the Lyve Pilot Bucket, and not to other data.",
+            "register-btn": "Register",
+            "clear": "Clear",
+            "s3AccountKey": "S3 Account Access Key",
+            "s3AccountSecretKey": "S3 Account Secret Key",
+            "userAccessKey": "IAM User Access Key",
+            "userAccessSecretKey": "IAM User Secret Key",
+            "created": "Access key created",
+            "access_key": "Access key",
+            "secret_key": "Secret key",
+            "iam_access_key": "IAM Access key",
+            "iam_secret_key": "IAM Secret key",
+            "bucket_name":"Bucket Name"
+        }
+    }
+}
